+++ packages/react-workspace/react-client-app/src/components/AnchorSection.tsx
@@ -1,0 +1,10 @@
+import { NimbleAnchor } from '@ni/nimble-react/anchor';
+import { SubContainer } from './SubContainer';
+
+export function AnchorSection(): React.JSX.Element {
+    return (
+        <SubContainer label="Anchor">
+            <div><NimbleAnchor href="#" appearance="prominent">Site root</NimbleAnchor></div>
+        </SubContainer>
+    );
+}
