+++ packages/react-workspace/react-client-app/src/components/CheckboxSection.tsx
@@ -1,0 +1,12 @@
+import { NimbleCheckbox } from '@ni/nimble-react/checkbox';
+import { SubContainer } from './SubContainer';
+
+export function CheckboxSection(): React.JSX.Element {
+    return (
+        <SubContainer label="Checkbox">
+            <NimbleCheckbox>Checkbox label</NimbleCheckbox>
+            <NimbleCheckbox>Checkbox label</NimbleCheckbox>
+            <NimbleCheckbox>Checkbox label</NimbleCheckbox>
+        </SubContainer>
+    );
+}
